Revert "Revert "ProvidedDependency""
This reverts commit 1bb15e337f997d21b6135cd03eb926d6219ce8e8.
